reafactor table into separate components
