+++ src/Components/Table/EnhancedTableHead/EnhancedTableHeader.const.js
@@ -1,0 +1,27 @@
+export const HeadCells = [
+  {
+    id: "name",
+    numeric: false,
+    disablePadding: true,
+    label: "",
+  },
+  {
+    id: "type",
+    numeric: true,
+    disablePadding: false,
+    label: "Type of Cuisine",
+  },
+  { id: "rating", numeric: true, disablePadding: false, label: "Rating 0-10" },
+  {
+    id: "cal",
+    numeric: true,
+    disablePadding: false,
+    label: "Calories Per Serving",
+  },
+  {
+    id: "cost",
+    numeric: true,
+    disablePadding: false,
+    label: "Cost Per Serving",
+  },
+];
